Revert "Fixes required for errorless MinGW-w64 compilation are done."
This reverts commit 9cad5d54cf0d78e92b4ee831299d83b657f42f2e as it
breaks compressing the meta folder (Windows 7, VS 2010).

+++ src/libs/7zip/win/CPP/7zip/Archive/Common/ParseProperties.cpp
@@ -50,7 +50,7 @@ HRESULT ParsePropDictionaryValue(const UString &srcStringSpec, UInt32 &dicSize)
     return E_INVALIDARG;
   if (srcString.Length() == numDigits)
   {
-    if (number >= (unsigned int)kLogarithmicSizeLimit) // PQR for MinGW-w64: Signed < Unsigned comparison.
+    if (number >= kLogarithmicSizeLimit)
       return E_INVALIDARG;
     dicSize = (UInt32)1 << (int)number;
     return S_OK;
